+++ demos/gallery/samples/widget/TextFontSize.js
@@ -1,0 +1,8 @@
+import { TextBox } from "@hpcc-js/common";
+
+new TextBox()
+    .target("target")
+    .text("Hello\nand\nWelcome!")
+    .fontSize(28)
+    .render()
+    ;