+++ packages/phosphor/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.1.6"></a>
+## [2.1.6](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/phosphor@2.1.5...@hpcc-js/phosphor@2.1.6) (2018-10-30)
+
+**Note:** Version bump only for package @hpcc-js/phosphor
+
+
+
+
+
 <a name="2.1.5"></a>
 ## [2.1.5](https://github.com/hpcc-systems/Visualization/compare/@hpcc-js/phosphor@2.1.4...@hpcc-js/phosphor@2.1.5) (2018-10-15)